Splitted functions creation into several classes, in an attempt to simplify each function's undestanding

+++ src/main/java/dridco/maven/GenerateIntegrationTestSourcesMojo.java
@@ -177,9 +177,8 @@ public class GenerateIntegrationTestSourcesMojo extends AbstractMojo {
 					.normalize() + ".java";
 			File targetPath = new File(javaTestsLocation + pkg);
 			File targetFile = new File(targetPath, targetName);
-			if (!targetPath.mkdirs()) {
-				throw new MojoExecutionException(
-						"Could not create required directories");
+			if (!(targetPath.exists() || targetPath.mkdirs())) {
+				throw new IOException("Could not create required directory " + targetPath);
 			}
 			IOUtils.write(java, new FileOutputStream(targetFile),
 					javaTestsEnconding);
